学食の投稿機能を実装

+++ src/main/java/com/example/Test07/repository/C8_Cafeteria/CafeteriaDAO.java
@@ -32,7 +32,7 @@ public class CafeteriaDAO {
                 .map((Map<String,Object> row) -> new CafeteriaPost(
                         row.get("postId").toString(),
                         row.get("menuId").toString(),
-                        (double)row.get("evaluation"),
+                        (int)row.get("evaluation"),
                         row.get("comment").toString()
                 )).toList();
